Wrap general XML-RPC calls to avoid url-http auth.
When calling 'xml-rpc-method-call', if the server responses with 401
Unauthorized, then 'url-http' will call 'url-auth' methods to try to
authenticate, using a different authentication mechanism that the one
already set. So that, we will wrap the 'xml-rpc-method-call' to nullify
the 'url-get-authentication' function on general calls. Moreover, if our
call gets unauthorized (e.g. server discards login cookies), we will try
to log in again and give the query a second chance.